py-packages/engine/pyproject.toml
[tool.poetry]
authors = ["Daven Quinn <dev@davenquinn.com>"]
description = "An API server application for the Corelle plate-rotation system"
license = "MIT"
name = "corelle.rotate"
version = "1.0.0"

packages = [
  {include = "corelle"},
]

[tool.poetry.dependencies]

Fiona = "^1.8.22"
IPython = "^8"
PyYAML = "^6.0"
"corelle.client" = {path = "../client", develop = true}
"macrostrat.database" = "2.0.0.dev3"
"macrostrat.utils" = "^1.0.1"
numba = "^0.56.3"
numpy = "^1.23.4"
numpy-quaternion = "^2022.4.2"
psycopg2-binary = "^2.9.5"
python = "^3.8"
rich = "^12.6.0"
typer = "^0.6.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
corelle = 'corelle.rotate.cli:cli'

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
